usbclerk: Windows service for signing and installing usb device drivers
When usbclerk service is up, any application can connect its named pipe with
a USB device (pid,vid) to sign and install. When done, the service will reply
with an ack or error code. Currently winusb driver is installed, in order to
use libusb for communicating with the device. The usbclerk service can be
installed/uninstalled by "usbclerk install"/"usbclerk uninstall".
usbclerktest is a sample cli app communicating with the usbclerk service.